# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CNC Worklist Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginBody {
  username: string;
  pin: string;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  admin: "admin",
  operator: "operator",
} as const;

export interface AuthUser {
  id: number;
  username: string;
  role: AuthUserRole;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  operator: "operator",
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
  active: boolean;
  createdAt: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  operator: "operator",
} as const;

export interface CreateUserBody {
  username: string;
  pin: string;
  role: CreateUserBodyRole;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  operator: "operator",
} as const;

export interface UpdateUserBody {
  /** @nullable */
  pin?: string | null;
  role?: UpdateUserBodyRole;
  active?: boolean;
}

export interface Project {
  projectId: string;
  projectNumber: string;
  address: string;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface Cutlist {
  cutlistId: string;
  pid: string;
  item: string;
  /** @nullable */
  lister?: string | null;
  /** @nullable */
  dateListed?: string | null;
  projectId: string;
}

export interface StockLevel {
  pcode: string;
  description: string;
  quantity: number;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  location?: string | null;
}

export interface Material {
  id: number;
  pcode: string;
  displayName: string;
  /** @nullable */
  notes?: string | null;
  isFavourite: boolean;
  createdAt: string;
}

export interface CreateMaterialBody {
  pcode: string;
  displayName: string;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateMaterialBody {
  pcode?: string;
  displayName?: string;
  /** @nullable */
  notes?: string | null;
}

export interface FavouriteResponse {
  isFavourite: boolean;
  materialId: number;
}

export type WorklistSummaryMachineType =
  (typeof WorklistSummaryMachineType)[keyof typeof WorklistSummaryMachineType];

export const WorklistSummaryMachineType = {
  B: "B",
  C: "C",
} as const;

export type WorklistSummaryStatus =
  (typeof WorklistSummaryStatus)[keyof typeof WorklistSummaryStatus];

export const WorklistSummaryStatus = {
  draft: "draft",
  active: "active",
  complete: "complete",
} as const;

export interface WorklistSummary {
  id: number;
  worklistNumber: string;
  projectId: string;
  projectNumber: string;
  projectAddress: string;
  machineType: WorklistSummaryMachineType;
  folderNumber: number;
  folderRef: string;
  status: WorklistSummaryStatus;
  itemCount: number;
  createdAt: string;
  /** @nullable */
  createdByUsername?: string | null;
}

export type WorklistMachineType =
  (typeof WorklistMachineType)[keyof typeof WorklistMachineType];

export const WorklistMachineType = {
  B: "B",
  C: "C",
} as const;

export type WorklistStatus =
  (typeof WorklistStatus)[keyof typeof WorklistStatus];

export const WorklistStatus = {
  draft: "draft",
  active: "active",
  complete: "complete",
} as const;

export interface WorklistItem {
  id: number;
  worklistId: number;
  /** @nullable */
  materialId?: number | null;
  pcode: string;
  displayName: string;
  quantity: number;
  /** @nullable */
  length?: number | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  notes?: string | null;
  sortOrder: number;
}

export interface Worklist {
  id: number;
  worklistNumber: string;
  projectId: string;
  projectNumber: string;
  projectAddress: string;
  cutlistRefs: string[];
  machineType: WorklistMachineType;
  folderNumber: number;
  folderRef: string;
  status: WorklistStatus;
  createdAt: string;
  /** @nullable */
  createdByUsername?: string | null;
  items: WorklistItem[];
}

export type CreateWorklistBodyMachineType =
  (typeof CreateWorklistBodyMachineType)[keyof typeof CreateWorklistBodyMachineType];

export const CreateWorklistBodyMachineType = {
  B: "B",
  C: "C",
} as const;

export interface CreateWorklistBody {
  projectId: string;
  projectNumber: string;
  projectAddress: string;
  cutlistRefs: string[];
  machineType: CreateWorklistBodyMachineType;
}

export type UpdateWorklistBodyStatus =
  (typeof UpdateWorklistBodyStatus)[keyof typeof UpdateWorklistBodyStatus];

export const UpdateWorklistBodyStatus = {
  draft: "draft",
  active: "active",
  complete: "complete",
} as const;

export interface UpdateWorklistBody {
  status?: UpdateWorklistBodyStatus;
  cutlistRefs?: string[];
}

export interface CreateWorklistItemBody {
  /** @nullable */
  materialId?: number | null;
  pcode: string;
  displayName: string;
  quantity: number;
  /** @nullable */
  length?: number | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateWorklistItemBody {
  quantity?: number;
  /** @nullable */
  length?: number | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  notes?: string | null;
  sortOrder?: number;
}

export interface WorklistStats {
  totalWorklists: number;
  draftCount: number;
  activeCount: number;
  completeCount: number;
  totalItemsThisMonth: number;
  nextWorklistNumber: string;
}

export interface FolderRef {
  machine: string;
  folderNumber: number;
  folderRef: string;
}

export interface AppSettings {
  filemakerServerUrl: string;
  filemakerDatabase: string;
  filemakerUsername: string;
  csvServerPath: string;
  worklistStartNumber: number;
  nextWorklistNumber: number;
}

export interface UpdateSettingsBody {
  filemakerServerUrl?: string;
  filemakerDatabase?: string;
  filemakerUsername?: string;
  filemakerPassword?: string;
  csvServerPath?: string;
  worklistStartNumber?: number;
}

export type ListProjectsParams = {
  search?: string;
};

export type ListCutlistsParams = {
  projectId?: string;
};

export type ListMaterialsParams = {
  search?: string;
  favouritesOnly?: boolean;
};

export type ListWorklistsParams = {
  status?: ListWorklistsStatus;
};

export type ListWorklistsStatus =
  (typeof ListWorklistsStatus)[keyof typeof ListWorklistsStatus];

export const ListWorklistsStatus = {
  draft: "draft",
  active: "active",
  complete: "complete",
} as const;

export type GetNextFolderParams = {
  machine: GetNextFolderMachine;
};

export type GetNextFolderMachine =
  (typeof GetNextFolderMachine)[keyof typeof GetNextFolderMachine];

export const GetNextFolderMachine = {
  B: "B",
  C: "C",
} as const;