# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CNC Worklist Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects from FileMaker
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project from FileMaker
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cutlists:
    get:
      operationId: listCutlists
      tags: [cutlists]
      summary: List cutlists from FileMaker
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of cutlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cutlist"
  /cutlists/{cutlistId}:
    get:
      operationId: getCutlist
      tags: [cutlists]
      summary: Get a single cutlist from FileMaker
      parameters:
        - name: cutlistId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cutlist details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cutlist"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stockbook/{pcode}:
    get:
      operationId: getStockLevel
      tags: [stockbook]
      summary: Get stock level for a PCODE from FileMaker
      parameters:
        - name: pcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock level info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockLevel"
        "404":
          description: PCODE not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List internal materials
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: favouritesOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create a new material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaterialBody"
      responses:
        "201":
          description: Material created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      summary: Get a material by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMaterial
      tags: [materials]
      summary: Update a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMaterialBody"
      responses:
        "200":
          description: Material updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /materials/{id}/stock:
    get:
      operationId: getMaterialStock
      tags: [materials]
      summary: Get live stock level for a material via its PCODE
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stock level
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockLevel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        pin:
          type: string
      required:
        - username
        - pin
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, operator]
      required:
        - id
        - username
        - role
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, operator]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - role
        - active
        - createdAt
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        pin:
          type: string
        role:
          type: string
          enum: [admin, operator]
      required:
        - username
        - pin
        - role
    UpdateUserBody:
      type: object
      properties:
        pin:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, operator]
        active:
          type: boolean
    Project:
      type: object
      properties:
        projectId:
          type: string
        projectNumber:
          type: string
        address:
          type: string
        date:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
      required:
        - projectId
        - projectNumber
        - address
    Cutlist:
      type: object
      properties:
        cutlistId:
          type: string
        pid:
          type: string
        item:
          type: string
        lister:
          type: ["string", "null"]
        dateListed:
          type: ["string", "null"]
        projectId:
          type: string
      required:
        - cutlistId
        - pid
        - item
        - projectId
    StockLevel:
      type: object
      properties:
        pcode:
          type: string
        description:
          type: string
        quantity:
          type: number
        unit:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
      required:
        - pcode
        - description
        - quantity
    Material:
      type: object
      properties:
        id:
          type: integer
        pcode:
          type: string
        displayName:
          type: string
        notes:
          type: ["string", "null"]
        isFavourite:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - pcode
        - displayName
        - isFavourite
        - createdAt
    CreateMaterialBody:
      type: object
      properties:
        pcode:
          type: string
        displayName:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - pcode
        - displayName
    UpdateMaterialBody:
      type: object
      properties:
        pcode:
          type: string
        displayName:
          type: string
        notes:
          type: ["string", "null"]
...